move simd intrinsics to their own file
